Software/1-Inference-LLM/requirements.txt
torch
torchvision
torchaudio
transformers
opencv-python-headless
gradio
sentencepiece
huggingface_hub
requests
groq
python-dotenv